requirements.txt
aiogram==2.19
aiohttp==3.8.1
aiosignal==1.2.0
aioredis==2.0.1
async-timeout==4.0.2
attrs==21.4.0
Babel==2.9.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
Flask==2.1.1
frozenlist==1.3.0
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
multidict==6.0.2
pyTelegramBotAPI==4.4.0
pytz==2022.1
# psycopg2==2.8.6
psycopg2-binary==2.8.6
requests==2.27.1
urllib3==1.26.9
Werkzeug==2.1.1
wincertstore==0.2
yarl==1.7.2
zipp==3.8.0
environs==9.4.0

sqlalchemy~=1.4.32